complete draft SYS-REQs 4-8 (#13)
User requirements (3,4) have been adjusted to BU characteristics.

Specific requirements for the reduction of difficulty and the recovery of difficulty retargeting have been pushed to lower level (pending software requirements).

Initial projected supportive hashpower will need to be estimated to arrive at a suitable setting for the difficulty value when it is reset to a lower value at fork time. Various community inputs will have to be considered closer to the time of release.