split up fork activation state tracking section, add description of configuration file to persist the state across restarts (MVHF-BU-DES-TRIG-10)
On fork activation, a new configuration file (btcfork.conf) will be written in the datadir.

The existence of this file will serve as an indicator to the software that the fork has already been activated, and that certain actions shall accordingly be performed in a way suitable to the forked state, e.g. connect to the new forked network instead of the old network.

The idea is that the user will be able to integrate (manually) some necessary settings from the btcfork.conf file into their bitcoin.conf file at some stage after fork activation, and then be able to discard the btcfork.conf entirely.

The software will not automatically modify the user's bitcoin.conf as this may be used by other clients, so the decision how to integrate any parameter changes is left up to the user.
At this stage, only a change of port number is expected to be necessary.