update with SW-reqs for signature change (*-REQ-9-*)
Includes specification of a grace period and mempool purge prior to expiry of this grace period.
Both these features still need to be implemented.

The grace period is made configurable with a possible minimum value of 0, because testing needs to show whether that is even feasible.
In practice it seems likely that a non-zero grace period would make sense, though this has also not been finalized.
The specification is thus written with extra flexibility here to permit testing around these requirements to shed light on what is possible / sensible.